moved executeCommand to seperate package as exported method
